requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3
Werkzeug==3.1.3
Pillow==12.0.0
